Make compile without C++11 (must have space between template opening '<' and namespace specifyer '::'). Fixed warning.
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@22088 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/solver_nonlin_nox/nox_nln_statustest_activeset.cpp
@@ -82,7 +82,7 @@ NOX::StatusTest::StatusType NOX::NLN::StatusTest::ActiveSet::checkStatus(
         // add a new map a the beginning of the deque
         cycling_maps_.push_front(cnlngrp->GetOldActiveSetMap(qtype_));
         // remove the last entry of the deque, if the max_cycle_size_ is exceeded
-        if (cycling_maps_.size()>max_cycle_size_)
+        if (cycling_maps_.size()>static_cast<std::size_t>(max_cycle_size_))
           cycling_maps_.pop_back();
 
         // check for cycling, if the set is not converged
